+++ mat3.h
@@ -7,21 +7,21 @@ public:
     mat3();
     ~mat3();
     mat3(int i_max_, int j_max_, int k_max_);
-    mat3(int i_max_, int j_max_, int k_max_, long double val);
-    void inic(long double val);
-    void set(int i, int j, int k, long double val);
+    mat3(int i_max_, int j_max_, int k_max_, double val);
+    void inic(double val);
+    void set(int i, int j, int k, double val);
     void setAll(mat3 &B);
-    void add(int i, int j, int k, long double val);
-    long double at(int i, int j, int k);
-    long double cols();
-    long double rows();
-    long double dims();
+    void add(int i, int j, int k, double val);
+    double at(int i, int j, int k);
+    double cols();
+    double rows();
+    double dims();
     void print();
 
 
 
     private:
-    long double * data;
+    double * data;
     int i_max_;
     int j_max_;
     int k_max_;
@@ -44,22 +44,22 @@ public:
     }
 
 
-    mat3::mat3(int i_max_, int j_max_, int k_max_, long double val) {
+    mat3::mat3(int i_max_, int j_max_, int k_max_, double val) {
         this->i_max_ = i_max_;
         this->j_max_ = j_max_;
         this->k_max_ = k_max_;
         inic(val);
     }
 
-    void mat3::inic(long double val) {
+    void mat3::inic(double val) {
         int n = i_max_ * j_max_ * k_max_;
-        data = new long double[n];
+        data = new double[n];
 
         for (int k = 0; k < n; k++)
             data[k] = val;
     }
 
-    void mat3::set(int i, int j, int k, long double val) {
+    void mat3::set(int i, int j, int k, double val) {
         data[k_max_ * j_max_ * i + j*k_max_ + k] = val;
     }
 
@@ -74,27 +74,27 @@ public:
     }
 
 
-    void mat3::add(int i, int j, int k, long double val) {
+    void mat3::add(int i, int j, int k, double val) {
         data[i_max_ * j_max_ * k + i * j_max_ + j] += val;
     }
 
 
-    long double mat3::at(int i, int j, int k) {
+    double mat3::at(int i, int j, int k) {
         return data[k_max_ * j_max_ * i + j*k_max_ + k];
     }
 
 
-    long double mat3::cols() {
+    double mat3::cols() {
         return j_max_;
     }
 
 
-    long double mat3::rows() {
+    double mat3::rows() {
         return i_max_;
     }
 
 
-    long double mat3::dims() {
+    double mat3::dims() {
         return k_max_;
     }
